+++ configs/databaseconfig.nix
@@ -1,4 +1,0 @@
-{ name, dbUser, dbUserPassword, init ? null }:
-{
-  inherit name dbUser dbUserPassword init;
-}